move `encrypt_bytes_stream_mode` to parent
