+++ src/crypto.rs
@@ -1,4 +1,3 @@
 pub mod decrypt;
-pub mod encrypt;
 pub mod key;
 pub mod primitives;
